Backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
alembic==1.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
